requirements.txt
crewai
streamlit
python-dotenv
fpdf2
litellm